requirements.txt
openai==1.99.9
together==1.5.24
transformers==4.55.2
tqdm==4.67.1
huggingface-hub==0.34.4
python-dotenv==1.1.0

# For local models
# torch==2.7.1+cu126
# accelerate==1.7.0
# bitsandbytes==0.46.0
# flash-attn==2.3.0  # For faster attention computation
